+++ R/model-zelig.R
@@ -257,7 +257,8 @@ z$methods(
     # Without dots for single and multiple equations
     temp_formula <- as.Formula(formula)
     if (sum(length(temp_formula)) <= 2)
-        .self$formula <- as.Formula(terms(temp_formula, data = localdata))
+        .self$formula <- as.Formula(terms(temp_formula,
+                                    data = localdata))
     else if (sum(length(temp_formula)) > 2) {
         f_dots <- attr(terms(temp_formula, data = localdata), "Formula_without_dot")
         if (!is.null(f_dots))
@@ -284,10 +285,10 @@ z$methods(
             if (.self$name == 'ivreg')
                 stop('logging values in the zelig call is not currently supported for ivreg models.',
                      call. = FALSE)
-            localformula <- transformer(formula, data = localdata, FUN = 'log',
-                                        f_out = TRUE)
-            localdata <- transformer(formula, data = localdata, FUN = 'log',
-                                     d_out = TRUE)
+            localformula <- transformer(formula, data = localdata,
+                                        FUN = 'log', f_out = TRUE)
+            localdata <- transformer(formula, data = localdata,
+                                     FUN = 'log', d_out = TRUE)
             .self$formula <- localformula
             .self$data <- localdata
         }
@@ -517,11 +518,12 @@ z$methods(
     s <- list(...)
 
     # This eliminates warning messages when factor rhs passed to lm() model in reduce() utility function
-    if(.self$category=="multinomial"){  # Perhaps find more robust way to test if dep.var. is factor
+    if(.self$category == "multinomial"){  # Perhaps find more robust way to test if dep.var. is factor
       f2 <- update(.self$formula, as.numeric(.) ~ .)
     } else {
       f2 <- .self$formula
     }
+
     f <- update(.self$formula, 1 ~ .)
     # update <- na.omit(.self$data) %>% # remove missing values
 
@@ -607,7 +609,8 @@ z$methods(
 )
 
 z$methods(
-  param = function(z.out, method="mvn") {
+  param = function(z.out, method = "mvn") {
+browser()
     if(identical(method,"mvn")){
       return(mvrnorm(.self$num, coef(z.out), vcov(z.out)))
     } else if(identical(method,"point")){