+++ NEWS.md
@@ -8,6 +8,11 @@ Zelig version 5.1-2
 
 ## Major changes
 
+-   New [EXPERIMENTAL] interface function `to_zelig` allows users to convert fitted model
+objects fitted outside of Zelig to a Zelig object. The function is called 
+within the `setx` wrapper if a non-Zelig object is supplied. Currently 
+only works for models fitted with `lm`. #189
+
 -   `get_se` and `get_pvalue` function wrappers created for `get_se` and
 `get_pvalue` methods, respectively. #269